+++ project-base/src/Shopsys/ShopBundle/Model/Category/CategoryRepository.php
@@ -177,11 +177,16 @@ class CategoryRepository extends NestedTreeRepository
      */
     public function findById($categoryId)
     {
-        return $this->getAllQueryBuilder()
-            ->andWhere('c.id = :categoryId')
-            ->setParameter('categoryId', $categoryId)
-            ->getQuery()
-            ->getOneOrNullResult();
+        $category = $this->getCategoryRepository()->find($categoryId);
+        /* @var $category \Shopsys\ShopBundle\Model\Category\Category */
+
+        if ($category->getParent() === null) {
+            // Copies logic from getAllQueryBuilder() - excludes root category
+            // Query builder is not used to be able to get the category from identity map if it was loaded previously
+            return null;
+        }
+
+        return $category;
     }
 
     /**